Move common functions within the files for specific ansatze
