+++ qadence/constructors/hea.py
@@ -1,13 +1,14 @@
 from __future__ import annotations
 
-from typing import Any, Type
+import itertools
+from typing import Any, Type, Union
 
-from qadence.blocks import AbstractBlock, block_is_qubit_hamiltonian, chain, tag
-from qadence.operations import CNOT, CPHASE, CRX, CRY, CRZ, CZ, RX, RY
+from qadence.blocks import AbstractBlock, block_is_qubit_hamiltonian, chain, kron, tag
+from qadence.constructors.hamiltonians import hamiltonian_factory
+from qadence.operations import CNOT, CPHASE, CRX, CRY, CRZ, CZ, RX, RY, HamEvo
 from qadence.types import Interaction, Strategy
 
-from .ansatze import DigitalEntanglers, _entanglers_analog, _entanglers_digital, _rotations_digital
-from .hamiltonians import hamiltonian_factory
+DigitalEntanglers = Union[CNOT, CZ, CRZ, CRY, CRX]
 
 
 def hea(
@@ -91,6 +92,112 @@ def hea(
 #############
 
 
+def _rotations_digital(
+    n_qubits: int,
+    depth: int,
+    param_prefix: str = "theta",
+    support: tuple[int, ...] | None = None,
+    operations: list[Type[AbstractBlock]] = [RX, RY, RX],
+) -> list[AbstractBlock]:
+    """Creates the layers of single qubit rotations in an HEA."""
+    if support is None:
+        support = tuple(range(n_qubits))
+    iterator = itertools.count()
+    rot_list: list[AbstractBlock] = []
+    for d in range(depth):
+        rots = [
+            kron(
+                gate(support[n], param_prefix + f"_{next(iterator)}")  # type: ignore [arg-type]
+                for n in range(n_qubits)
+            )
+            for gate in operations
+        ]
+        rot_list.append(chain(*rots))
+    return rot_list
+
+
+def _entangler(
+    control: int,
+    target: int,
+    param_str: str,
+    op: Type[DigitalEntanglers] = CNOT,
+) -> AbstractBlock:
+    """
+    Create a 2-qubit operation for the digital HEA.
+
+    Args:
+        control (int): control qubit index
+        target (int): target qubit index
+        param_str (str): base for naming the variational parameter if parametric block
+        op (Type[DigitalEntanglers]): 2-qubit operation (CNOT, CZ, CRX, CRY, CRZ or CPHASE)
+
+    Returns:
+        A 2-qubit operation block
+    """
+    if op in [CNOT, CZ]:
+        return op(control, target)  # type: ignore
+    elif op in [CRZ, CRY, CRX, CPHASE]:
+        return op(control, target, param_str)  # type: ignore
+    else:
+        raise ValueError("Provided entangler not accepted for digital ansatz")
+
+
+def _entanglers_digital(
+    n_qubits: int,
+    depth: int,
+    param_prefix: str = "theta",
+    support: tuple[int, ...] | None = None,
+    periodic: bool = False,
+    entangler: Type[DigitalEntanglers] = CNOT,
+) -> list[AbstractBlock]:
+    """Creates the layers of digital entangling operations in an HEA.
+
+    Args:
+        n_qubits (int): number of qubits in the block.
+        depth (int): number of layers of the HEA.
+        param_prefix (str): the base name of the variational parameters
+        support (tuple): qubit indexes where the HEA is applied.
+        periodic (bool): if the qubits should be linked periodically.
+            periodic=False is not supported in emu-c.
+        entangler (AbstractBlock): 2-qubit entangling operation.
+            Supports CNOT, CZ, CRX, CRY, CRZ. Controlld rotations
+            will have variational parameters on the rotation angles.
+    """
+    if support is None:
+        support = tuple(range(n_qubits))
+    iterator = itertools.count()
+    ent_list: list[AbstractBlock] = []
+    for d in range(depth):
+        ents = []
+        ents.append(
+            kron(
+                _entangler(
+                    control=support[n],
+                    target=support[n + 1],
+                    param_str=param_prefix + f"_ent_{next(iterator)}",
+                    op=entangler,
+                )
+                for n in range(n_qubits)
+                if not n % 2 and n < n_qubits - 1
+            )
+        )
+        if n_qubits > 2:
+            ents.append(
+                kron(
+                    _entangler(
+                        control=support[n],
+                        target=support[(n + 1) % n_qubits],
+                        param_str=param_prefix + f"_ent_{next(iterator)}",
+                        op=entangler,
+                    )
+                    for n in range(n_qubits - (not periodic))
+                    if n % 2
+                )
+            )
+        ent_list.append(chain(*ents))
+    return ent_list
+
+
 def hea_digital(
     n_qubits: int,
     depth: int = 1,
@@ -151,6 +258,14 @@ def hea_digital(
 ###########
 ## sDAQC ##
 ###########
+
+
+def _entanglers_analog(
+    depth: int,
+    param_prefix: str = "theta",
+    entangler: AbstractBlock | None = None,
+) -> list[AbstractBlock]:
+    return [HamEvo(entangler, param_prefix + f"_t_{d}") for d in range(depth)]  # type: ignore
 
 
 def hea_sDAQC(
